+++ docs_snippets/lib/snippets/getting_started/getting_started.dart
@@ -29,7 +29,7 @@ class Application extends StatelessWidget {
       // There is a known issue with implicitly animated widgets where their transition occurs AFTER the theme's.
       // See https://github.com/duobaseio/forui/issues/670.
       theme: theme.toApproximateMaterialTheme(),
-      builder: (_, child) => FAnimatedTheme(
+      builder: (_, child) => FTheme(
         data: theme,
         child: FToaster(child: child!),
       ),