+++ docs_snippets/lib/snippets/getting_started/material.dart
@@ -15,7 +15,7 @@ class Application extends StatelessWidget {
   // {@highlight}
   Widget build(BuildContext context) => CupertinoApp(
     // {@endhighlight}
-    builder: (context, child) => FAnimatedTheme(data: FThemes.zinc.light, child: child!),
+    builder: (context, child) => FTheme(data: FThemes.zinc.light, child: child!),
     home: const FScaffold(child: Placeholder()),
   );
 }