+++ docs_snippets/lib/snippets/localization/default.dart
@@ -14,7 +14,7 @@ final app =
         // Add locales supported by your application here.
         // {@endhighlight}
       ],
-      builder: (context, child) => FAnimatedTheme(data: FThemes.zinc.light, child: child!),
+      builder: (context, child) => FTheme(data: FThemes.zinc.light, child: child!),
       home: const FScaffold(child: Placeholder()),
     );
 // {@endsnippet}