+++ docs_snippets/lib/snippets/localization/localization.dart
@@ -8,7 +8,7 @@ final app =
       localizationsDelegates: FLocalizations.localizationsDelegates,
       supportedLocales: FLocalizations.supportedLocales,
       // {@endhighlight}
-      builder: (context, child) => FAnimatedTheme(data: FThemes.zinc.light, child: child!),
+      builder: (context, child) => FTheme(data: FThemes.zinc.light, child: child!),
       home: const FScaffold(child: Placeholder()),
     );
 // {@endsnippet}