+++ docs_snippets/lib/snippets/themes/customization/custom_font_family.dart
@@ -3,7 +3,7 @@ import 'package:forui/forui.dart';
 
 // {@snippet}
 @override
-Widget build(BuildContext context) => FAnimatedTheme(
+Widget build(BuildContext context) => FTheme(
   data: FThemeData(
     colors: FThemes.zinc.light.colors,
     // {@highlight}